GitHub/fffffw/fitness_bot/requirements.txt
python-telegram-bot==21.5
openai
flask
apscheduler